Fix running of tests with --pyargs
