+++ hyperion/conftest.py
@@ -1,23 +1,7 @@
 import warnings
 
-import pytest
-
 from .util.nans import NaNWarning
-
-
-def pytest_addoption(parser):
-    parser.addoption('--generate-reference', help="generate reference results for bit-level tests", type="string")
-    parser.addoption('--enable-bit-level-tests', help="enable bit-level tests", action="store_true")
 
 
 def pytest_configure(config):
     warnings.simplefilter('error', NaNWarning)
-
-
-def pytest_collection_modifyitems(config, items):
-    if config.getoption("--enable-bit-level-tests"):
-        return
-    skip_bit_level = pytest.mark.skip(reason="need --enable-bit-level-tests option to run")
-    for item in items:
-        if "bitlevel" in item.keywords:
-            item.add_marker(skip_bit_level)
