+++ setup.cfg
@@ -29,3 +29,7 @@ hyperion.model.tests = data/*.rtout, data/*.hdf5
 hyperion.importers.tests = data/*.hdf5
 hyperion.grid.tests = data/*.hdf5, data/DD0010/*
 hyperion.testing = coveragerc
+
+[options.entry_points]
+pytest11 =
+    hyperion = hyperion.testing.pytest_plugin
